backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.27.0
python-dotenv==1.0.1
PyJWT==2.10.1
websockets==13.1
